+++ R/desc.R
@@ -4,7 +4,7 @@
 ###############################################################################
 
 
-#' Constructor function for the desc object
+#' Constructor function for desc objects
 #' 
 #' @param output a data.frame object
 #' @param y Character indicating the response variable
@@ -27,7 +27,7 @@
 #' Default = "Statistics"
 #' @param regrouped Logical. used to indicate if the desc object regroup several desc objects (see \code{\link{regroup.desc}} )
 #' @param rbind.label Character. The name of the label in case several desc objects are regrouped
-#' @param title Character. The title by default of the desc object
+#' @param title Character. The title of the desc object that will be printed and used in \code{\link{report.doc}}
 #' 
 #' 
 #' @description