+++ R/report.doc.R
@@ -29,9 +29,9 @@
 #' @description
 #' \code{report.doc} 
 #' This function enables to export the table created with \code{\link{report.quali}} \code{\link{report.quanti}} or \code{\link{report.lsmeans}}
-#' to a Microsoft Word or a R markdown document in standrad easy to read format. 
+#' to a Microsoft Word or a R markdown document. 
 #' 
-#' It's also possible to use it to have a preview of the table in HTML format if the doc argument is NULL.
+#' It's also possible to use report.doc to have a preview of the table in HTML format if the \code{doc} argument is NULL.
 #' 
 #' @details
 #' 
@@ -190,7 +190,7 @@
 #' 
 #'anov1=Anova(mod1)
 #'
-#'tab.mod1=report.lsmeans(lsm=test1,at.row="TIMEPOINT")
+#'tab.mod1=report.lsmeans(lsm=test1)
 #'
 #'gg.mod1=plot(tab.mod1,title="LS-Means response evolution as a function of time\n
 #' by treatment group (95% CI)",
@@ -219,7 +219,7 @@
 #' 
 #'test3=emmeans(mod3,~GROUP|TIMEPOINT)
 #' 
-#'tab.mod3=report.lsmeans(lsm=test3,at.row="TIMEPOINT")
+#'tab.mod3=report.lsmeans(lsm=test3)
 #'
 #'gg.mod3=plot(tab.mod3,title="LS-Means response evolution as a function of time\n
 #'by treatment group (95% CI Mixed model)",
@@ -229,15 +229,17 @@
 #'
 #'contr=contrast(test3, "trt.vs.ctrl", ref = "A")
 #'
-#'tab.mod3.contr=report.lsmeans(lsm=contr,at="TIMEPOINT")
+#'tab.mod3.contr=report.lsmeans(lsm=contr)
 #'
 #'gg.mod3.contr=plot(tab.mod3.contr,title="LS-Means contrast versus reference A\n
 #'				(95% CI Mixed model)",
 #'		legend.label="Treatment groups",ylab="Y mean",add.ci=T,add.line=F)
 #'
 #'
-#'
+#' ############################################################
 #' # Generalized Logistic Linear model (order 2 interaction):
+#' ############################################################
+#' 
 #' # Anova LS-Means and graph reporting ##########
 #' 
 #'mod4=glm(y_logistic~baseline+GROUP+TIMEPOINT+GROUP*TIMEPOINT,
@@ -598,6 +600,7 @@ report.doc.desc=function(table,title=NULL,colspan.value=NULL,doc=NULL,
 }
 
 #' @param type.anova Passed to \code{Anova} function from car package (see its documentation).
+#' @param pretty.label Logical. Default to FALSE. If TRUE, use the function \code{make.label} with default option on the rownames of the anova table 
 #' 
 #' @importFrom xtable xtable
 #' @rdname report.doc
@@ -606,7 +609,7 @@ report.doc.desc=function(table,title=NULL,colspan.value=NULL,doc=NULL,
 
 
 report.doc.anova=function(table,title="Anova table",type.anova=3,doc=NULL,numbering=T,
-		init.numbering=F,font.name="Times",font.size=10,page.break=T,...)
+		init.numbering=F,font.name="Times",font.size=10,page.break=T,pretty.label=FALSE,...)
 {
 	
 	
@@ -634,6 +637,12 @@ report.doc.anova=function(table,title="Anova table",type.anova=3,doc=NULL,number
 	
 	ncol=ncol(as.data.frame(table))
 	
+	
+	if(pretty.label)
+	{
+		rownames(table)=make.label(rownames(table))
+	}
+	
 	xtab=xtable(table)
 	ft=xtable_to_flextable(xtab,NA.string = "-")
 	
@@ -658,8 +667,8 @@ report.doc.anova=function(table,title="Anova table",type.anova=3,doc=NULL,number
 	
 	# change font 
 	
-	ft=font(ft,fontname=font.name,part ="all")
-	ft=fontsize(ft,size=font.size,part ="all")
+	ft=flextable::font(ft,fontname=font.name,part ="all")
+	ft=flextable::fontsize(ft,size=font.size,part ="all")
 	
 	if(!is.null(doc))
 	{